refactor: getting rid of IMinecraft for global extensions
